requirements.txt
aiogram==3.12.0
python-dotenv==0.21.1
psycopg2
asyncpg
SQLAlchemy